delete React version, add getter for get state
