+++ 9_10lesson/src/App.js
@@ -1,37 +1,29 @@
 import React from 'react'
-import { BrowserRouter as Router, Switch, Route, NavLink } from 'react-router-dom'
-import Footer from './Footer'
-
-import ReduxApp from './AppRedux'
-import ReactApp from './AppReact'
-
 import { useDispatch, useSelector } from 'react-redux'
-import { changeTheme } from './redux/actions'
+import UsersList from './Components/UsersList'
+import { changeTheme, fetchUsers } from './redux/actions'
+import Footer from './Footer'
+import getTheme from './redux/selectors/themeSelector'
 
 function App() {
-  const theme = useSelector(state => state.theme.setTheme) //useSelector - маппинг значения из store.
-  document.body.className = theme
   const dispatch = useDispatch()
-
+  const theme = useSelector(state => getTheme(state))
+  document.body.className = theme
   const newTheme = document.body.classList.contains('light') ? 'dark' : 'light'
+  
   return (
-    <Router>
-      <div className="wrapper">
-        <div className="flex">
-          <nav>
-            <NavLink to="/">React</NavLink>
-            <NavLink to="/ReduxApp">Redux</NavLink>
-          </nav>
-          <button onClick={() => dispatch(changeTheme(newTheme))}>Change theme</button>
-        </div>
-
-        <Switch>
-          <Route exact path="/" component={ReactApp} />
-          <Route exact path="/ReduxApp" component={ReduxApp} />
-        </Switch>
-      </div>
+    <div className="wrapper">
+      <h1>User list from REDAX</h1>
+      <button className="changeTheme" onClick={() => dispatch(changeTheme(newTheme))}>
+        Change theme
+      </button>
+      <button className="addUser redux" onClick={() => dispatch(fetchUsers())}>
+        Add users
+      </button>
+      <UsersList />
       <Footer />
-    </Router>
+    </div>
   )
 }
+
 export default App
