+++ 9_10lesson/src/Footer.js
@@ -4,9 +4,8 @@ export default function footer() {
   return (
     <footer>
       <div className="wrapper footer">
-        <p>
-          ©2020<span>By AndruB</span>
-        </p>
+        <p>©2020</p>
+        <span>By AndruB</span>
       </div>
     </footer>
   )
