--- 9_10lesson/src/context.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const ContextDel = React.createContext()
-
-export default ContextDel
