+++ 9_10lesson/src/index.js
@@ -1,11 +1,10 @@
 import React from 'react'
 import { render } from 'react-dom'
-import { compose, createStore, applyMiddleware } from 'redux'
+import { createStore, compose, applyMiddleware } from 'redux'
 import { Provider } from 'react-redux'
 import thunk from 'redux-thunk'
 import './index.css'
 import App from './App'
-import * as serviceWorker from './serviceWorker'
 import { rootReducer } from './redux/rootReducer'
 
 const store = createStore(
@@ -15,6 +14,7 @@ const store = createStore(
     window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
   )
 )
+
 const app = (
   <Provider store={store}>
     <App />
@@ -22,8 +22,3 @@ const app = (
 )
 
 render(app, document.getElementById('root'))
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister()
