+++ 9_10lesson/src/redux/actions.js
@@ -3,8 +3,8 @@ import { DELL_USERS, FETCH_USERS, THEME } from './types'
 export function fetchUsers() {
   return async dispatch => {
     const response = await fetch('https://api.github.com/users')
-    const gitUsers = await response.json()
-    dispatch({ type: FETCH_USERS, payload: gitUsers }) //функция которая меняет состояние store
+    const getUsers = await response.json()
+    dispatch({ type: FETCH_USERS, payload: getUsers })
   }
 }
