+++ 9_10lesson/src/redux/selectors/themeSelector.js
@@ -1,0 +1,4 @@
+
+const getTheme = state => state.theme.setTheme
+
+export default getTheme
