+++ 9_10lesson/src/redux/selectors/userSelector.js
@@ -1,0 +1,4 @@
+
+const getUser = state => state.users.fetchedUsers
+
+export default getUser
