+++ 9_10lesson/src/redux/usersReducer.js
@@ -5,7 +5,6 @@ const initialState = {
 }
 
 export const usersReducer = (state = initialState, action) => {
-  console.log('action', action)
   switch (action.type) {
     case FETCH_USERS:
       return { ...state, fetchedUsers: action.payload }
